Add support for query parameters
Query parameters may now be passed to the builder. To provide more
flexibility, borrows in the Builder are now limited to the lifetime of
the builder with the exception of the Connection. This required a slight
change in API--the cursor is now created at the same time as a `Cursor`
is generated.